Create MLGameStateStack file
