+++ ocky/Observables/MLGame.swift
@@ -10,37 +10,6 @@ import SwiftUI
 import GameKit
 import Combine
 
-class MLGameStateStack {
-  var stack: [MLGameState]
-  
-  init() {
-    stack = []
-  }
-  
-  func push(_ state: MLGameState) {
-    if case .inQuestion = self.peek(),
-       case .inQuestion = state {
-      self.pop()
-    }
-    
-    switch state {
-      case .findMatch, .loadMatch, .loadMatches, .result:
-        return
-      default: stack.append(state)
-    }
-  }
-  
-  func pop() {
-    stack = stack.dropLast()
-    stack.count == 0 ? push(.idle) : ()
-  }
-  
-  func peek() -> MLGameState? {
-    stack.last
-  }
-}
-
-
 class MLGame: NSObject, ObservableObject {
   var gameData: MLGameData
   lazy var stateStack: MLGameStateStack = MLGameStateStack()
@@ -153,6 +122,7 @@ class MLGame: NSObject, ObservableObject {
     guard let data = gameData.encode() else { return }
     do {
       if activeMatch?.currentParticipant?.player?.displayName == GKLocalPlayer.local.displayName {
+        // filter out the user info
         try await activeMatch?.participantQuitInTurn(with: .quit,
                                                      nextParticipants: activeMatch?.participants ?? [],
                                                      turnTimeout: GKTurnTimeoutDefault, match: data)
